ai/ai-endpoints/audio-virtual-assistant/requirements.txt
openai==1.13.3
streamlit==1.36.0
streamlit-mic-recorder==1.16.0
nvidia-riva-client==2.15.1
python-dotenv==1.0.1
